Revert "Merge 6.12.39 into android16-6.12-lts"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2afc1558-54de-4ec5-9cf1-b3161ad08f0d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

+++ net/ipv4/tcp.c
@@ -1148,7 +1148,7 @@ restart:
 		goto do_error;
 
 	while (msg_data_left(msg)) {
-		int copy = 0;
+		ssize_t copy = 0;
 
 		skb = tcp_write_queue_tail(sk);
 		if (skb)
